구글 로그인 관련 코드를 리팩토링 이후 방식으로 재적용
