+++ app/src/main/java/com/wafflestudio/snutt2/data/user/UserRepositoryImpl.kt
@@ -13,7 +13,7 @@ import com.wafflestudio.snutt2.core.network.SNUTTNetworkDataSource
 import com.wafflestudio.snutt2.core.network.model.GetUserFacebookResults
 import com.wafflestudio.snutt2.core.network.model.PostAccessTokenByAuthCodeParams
 import com.wafflestudio.snutt2.core.network.model.PostSocialLoginParams
-import com.wafflestudio.snutt2.lib.network.SNUTTRestApiForGoogle
+import com.wafflestudio.snutt2.core.network.retrofit.google.SNUTTRestApiForGoogle
 import com.wafflestudio.snutt2.lib.network.dto.core.toDatabaseModel
 import com.wafflestudio.snutt2.lib.network.dto.core.toExternalModel
 import com.wafflestudio.snutt2.model.TableTrimParam
