--- app/src/main/java/com/wafflestudio/snutt2/lib/network/SNUTTRestApiForGoogle.kt
@@ -1,13 +1,0 @@
-package com.wafflestudio.snutt2.lib.network
-
-import com.wafflestudio.snutt2.lib.network.dto.PostAccessTokenByAuthCodeParams
-import com.wafflestudio.snutt2.lib.network.dto.PostAccessTokenByAuthCodeResults
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface SNUTTRestApiForGoogle {
-    @POST("/token")
-    suspend fun _getAccessTokenByAuthCode(
-        @Body body: PostAccessTokenByAuthCodeParams,
-    ): PostAccessTokenByAuthCodeResults
-}
