--- app/src/main/java/com/wafflestudio/snutt2/lib/network/dto/PostAccessTokenByAuthCodeParams.kt
@@ -1,13 +1,0 @@
-package com.wafflestudio.snutt2.lib.network.dto
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class PostAccessTokenByAuthCodeParams(
-    @Json(name = "code") val authCode: String,
-    @Json(name = "client_id") val clientId: String,
-    @Json(name = "client_secret") val clientSecret: String,
-    @Json(name = "redirect_uri") val redirectUri: String = "",
-    @Json(name = "grant_type") val grantType: String = "authorization_code",
-)
