+++ app/src/main/res/values/strings.xml
@@ -3,8 +3,6 @@
     <string name="privacy" translatable="false">privacy_policy</string>
     <string name="member" translatable="false">member</string>
 
-    <string name="api_google_server">https://oauth2.googleapis.com</string>
-
     <!-- For sign up fragment -->
     <string name="sign_up_terms_1">회원가입을 누르시면</string>
     <string name="sign_up_terms_2">일반 이용 약관</string>
