+++ core/network/src/main/java/com/wafflestudio/snutt2/core/network/SNUTTNetworkDataSourceForGoogle.kt
@@ -1,0 +1,10 @@
+package com.wafflestudio.snutt2.core.network
+
+import com.wafflestudio.snutt2.core.network.model.PostAccessTokenByAuthCodeParams
+import com.wafflestudio.snutt2.core.network.model.PostAccessTokenByAuthCodeResults
+
+interface SNUTTNetworkDataSourceForGoogle {
+    suspend fun _getAccessTokenByAuthCode(
+        body: PostAccessTokenByAuthCodeParams,
+    ): PostAccessTokenByAuthCodeResults
+}
