+++ core/network/src/main/java/com/wafflestudio/snutt2/core/network/model/PostAccessTokenByAuthCodeParams.kt
@@ -1,2 +1,13 @@
 package com.wafflestudio.snutt2.core.network.model
 
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class PostAccessTokenByAuthCodeParams(
+    @Json(name = "code") val authCode: String,
+    @Json(name = "client_id") val clientId: String,
+    @Json(name = "client_secret") val clientSecret: String,
+    @Json(name = "redirect_uri") val redirectUri: String = "",
+    @Json(name = "grant_type") val grantType: String = "authorization_code",
+)
