+++ core/network/src/main/res/values/strings.xml
@@ -1,2 +1,3 @@
 <resources>
+    <string name="api_google_server">https://oauth2.googleapis.com</string>
 </resources>
